Update new logic

+++ appsystem/Bola.php
@@ -27,13 +27,13 @@ try {
 //                        [8] => 2017-07-19 09:15:09
 //                        [9] => reg
 
-                    $checkQuery = "SELECT * FROM tb_sms_push WHERE "
+                    $checkQuery = "SELECT content_number FROM tb_push_summary WHERE "
                             . "telco = '$expldData[0]' AND "
                             . "shortcode = '$expldData[1]' AND "
                             . "msisdn = '$expldData[2]' AND "
                             . "keyword = '$expldData[3]' "
                             . "ORDER BY id_push DESC LIMIT 1";
-                    $result = $this->db->query($checkQuery);
+                    $result = $this->smspush->query($checkQuery);
                     $smsPushData = $result->fetchAll();
 
                     $sessionRand = rand(1, 99999999);
@@ -56,22 +56,23 @@ try {
                             . "id_app = '$key[id_app]'";
                     $appConfigResult = $this->db->query($appConfigQuery);
                     $appConfig = $appConfigResult->fetchAll()[0];
-                                        
+
                     //
                     if ($result->numRows() > 0) {
                         // Create App Message Only
-                        $contentSequence = $smsPushData[0]['content_number'] + 1;
-                        $contentQuery = "SELECT * FROM tb_apps_content WHERE "
-                                . "keyword = '$appName' AND "
-                                . "content_number = '$contentSequence'";
-                        $contentResult = $this->db->query($contentQuery);
-                        $cntn = $contentResult->fetchAll()[0];
+                        //$contentSequence = $smsPushData[0]['content_number'] + 1;
+                        //$contentQuery = "SELECT * FROM tb_apps_content WHERE "
+                        //        . "keyword = '$appName' AND "
+                        //        . "content_number = '$contentSequence'";
+                        //$contentResult = $this->db->query($contentQuery);
+                        //$cntn = $contentResult->fetchAll()[0];
 
                         // seqNumber Reply regType pull/pushType cost statusSend
                         //$contentApp = $expldData[0] . '|' . $expldData[1] . '|' . $expldData[2] . '|' . $expldData[3] . '|' . $expldData[4] . '||' . $expldData[6] . '|' . $expldData[7] . '|' . $expldData[8] . '|' . $expldData[9] . '|' . $cntn['content_number'] . '|' . $cntn['content_field'] . '|1|pull|' . $appConfig['cost_pull'] . '|1|push;iod;' . $appName . ';dailypush';
                         $contentApp = $dataFiles . '|0|Welcome Message Karena reg 2x|1|pull|0|1|reg;iod;' . $appName;
                         $fileApp = $pathPush . '/' . $sessionRand . 'dua.txt';
                     } else {
+                        $contentSequence = 1;
                         $contentWelcome = $dataFiles . '|0|Welcome Message Karena reg 1x|1|pull|0|1|reg;iod;' . $appName;
                         $fileWelcome = $pathPush . '/' . $expldData[5] . 'satu.txt';
                         // Create App File
@@ -84,7 +85,7 @@ try {
                                 // Create  App Message
                                 $contentQuery = "SELECT * FROM tb_apps_content WHERE "
                                         . "keyword = '$appName' AND "
-                                        . "content_number = '1'";
+                                        . "content_number = '$contentSequence'";
                                 $contentResult = $this->db->query($contentQuery);
                                 $cntn = $contentResult->fetchAll()[0];
 
@@ -96,15 +97,23 @@ try {
                     }
 
                     // Execute Create App File
-                    $createPush = fopen($fileApp, "w");
-                    if ($createPush) {
-                        $fwsMain = fwrite($createPush, $contentApp);
-                        if ($fwsMain) {
-                            if (unlink($appFolder . "/" . $fileNames)) {
-                                echo date('Y-m-d h:i:s') . " : Create Push & App File Unlink Success \n";
+                    $updateQuery = "UPDATE tb_members SET "
+                            . "content_seq = '$contentSequence' WHERE "
+                            . "shortcode = '$expldData[1]' AND "
+                            . "msisdn = '$expldData[2]' AND "
+                            . "app = '$expldData[3]'";
+                    $resultUpdate = $this->db->query($updateQuery);
+                    //
+                    if ($resultUpdate->numRows() > 0) {
+                        $createPush = fopen($fileApp, "w");
+                        if ($createPush) {
+                            $fwsMain = fwrite($createPush, $contentApp);
+                            if ($fwsMain) {
+                                if (unlink($appFolder . "/" . $fileNames)) {
+                                    echo date('Y-m-d h:i:s') . " : Create Push & App File Unlink Success \n";
+                                }
                             }
                         }
-                        //}
                     }
                 }
             }
